+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @hapergg/harness-comet-cli
 
+## 0.1.2
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @hapergg/harness-comet-adapter-memory@0.1.2
+  - @hapergg/harness-comet-adapter-playwright@0.1.2
+  - @hapergg/harness-comet-comet-adapter@0.1.2
+  - @hapergg/harness-comet-core@0.1.2
+  - @hapergg/harness-comet-schema@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes