+++ packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @hapergg/harness-comet-core
 
+## 0.1.2
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @hapergg/harness-comet-schema@0.1.2
+  - @hapergg/harness-comet-sdk@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes