+++ packages/playwright/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hapergg/harness-comet-playwright
 
+## 0.1.2
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @hapergg/harness-comet-schema@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes