+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hapergg/harness-comet-schema
 
+## 0.1.2
+
+### Patch Changes
+
+- update
+
 ## 0.1.1
 
 ### Patch Changes